requirements.txt
# Core dependencies
qdrant-client>=1.7.0
sentence-transformers>=2.3.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
mcp-sdk>=0.1.0
click>=8.1.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
markdown>=3.5.0
python-frontmatter>=1.1.0
gitpython>=3.1.0
rich>=13.7.0
structlog>=24.1.0
numpy>=1.26.0
scikit-learn>=1.4.0

# Development dependencies
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-aiohttp>=1.0.0
aioresponses>=0.7.0
black>=24.1.0
isort>=5.13.0
mypy>=1.8.0
ruff>=0.2.0
types-beautifulsoup4>=4.12.0
types-markdown>=3.5.0
types-aiofiles>=23.2.0

# Type stubs
types-python-dateutil>=2.8.0
types-requests>=2.31.0
types-setuptools>=69.0.0
types-PyYAML>=6.0.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.5.0
mkdocstrings>=0.24.0
mkdocstrings-python>=1.8.0

# Debugging
debugpy>=1.8.0
ipython>=8.21.0

# Testing utilities
coverage>=7.4.0
tox>=4.12.0
